refactor: migrate markdown renderer
