+++ multimodal/tarko/agent-ui/package.json
@@ -9,17 +9,9 @@
     "framer-motion": "12.19.2",
     "jotai": "^2.7.0",
     "jsonrepair": "3.12.0",
-    "katex": "^0.16.9",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^4.11.0",
-    "react-markdown": "10.1.0",
-    "react-syntax-highlighter": "15.6.1",
-    "rehype-highlight": "7.0.2",
-    "rehype-katex": "^7.0.1",
-    "remark-gfm": "4.0.1",
-    "remark-github-blockquote-alert": "1.3.0",
-    "remark-math": "^6.0.0",
     "socket.io-client": "^4.7.2",
     "uuid": "^9.0.0"
   },
@@ -35,7 +27,6 @@
     "@tarko/interface": "workspace:*",
     "@types/react": "^18.2.75",
     "@types/react-dom": "^18.2.24",
-    "@types/react-syntax-highlighter": "^15.5.11",
     "@types/uuid": "^9.0.8",
     "autoprefixer": "^10.4.20",
     "diff": "^8.0.2",
