+++ multimodal/tarko/agent-ui/src/standalone/navbar/Navbar.tsx
@@ -1,4 +1,5 @@
 import React, { useCallback, useEffect, useRef, useState } from 'react';
+import { useDarkMode } from '@tarko/ui';
 import { ShareButton } from '@/standalone/share';
 import { AboutModal } from './AboutModal';
 
@@ -23,22 +24,21 @@ import { GoSidebarCollapse, GoSidebarExpand } from 'react-icons/go';
 import { useLayout } from '@/common/hooks/useLayout';
 import { useSession } from '@/common/hooks/useSession';
 import { useReplayMode } from '@/common/hooks/useReplayMode';
-import { useDarkMode } from '@/common/hooks/useDarkMode';
 import { useLogoType } from '@/common/hooks/useLogoType';
 import { apiService } from '@/common/services/apiService';
 import { NavbarModelSelector } from './ModelSelector';
 import { getLogoUrl, getAgentTitle, getWorkspaceNavItems } from '@/config/web-ui-config';
 import type { WorkspaceNavItemIcon } from '@tarko/interface';
 import { getModelDisplayName } from '@/common/utils/modelUtils';
-import { 
-  Box, 
-  Typography, 
-  ThemeProvider, 
-  Menu, 
-  MenuItem, 
-  Divider, 
+import {
+  Box,
+  Typography,
+  ThemeProvider,
+  Menu,
+  MenuItem,
+  Divider,
   IconButton,
-  createBasicMuiTheme 
+  createBasicMuiTheme,
 } from '@tarko/ui';
 
 import './Navbar.css';
