+++ multimodal/tarko/ui/package.json
@@ -27,7 +27,14 @@
     "parse-diff": "0.11.1",
     "@monaco-editor/react": "4.7.0",
     "highlight.js": "^11.9.0",
-    "@mui/material": "^5.15.11"
+    "@mui/material": "^5.15.11",
+    "katex": "^0.16.9",
+    "react-markdown": "10.1.0",
+    "rehype-highlight": "7.0.2",
+    "rehype-katex": "^7.0.1",
+    "remark-gfm": "4.0.1",
+    "remark-github-blockquote-alert": "1.3.0",
+    "remark-math": "^6.0.0"
   },
   "devDependencies": {
     "@rslib/core": "0.10.0",
