+++ multimodal/tarko/ui/src/components/markdown-renderer/components/CodeBlock.tsx
@@ -20,7 +20,7 @@ export const CodeBlock: React.FC<CodeBlockProps> = ({ inline, className, childre
 
   return (
     <div className="my-2">
-      <pre 
+      <pre
         className="bg-[#f5f5f5] dark:bg-[#111111] dark:border-gray-700/50 rounded-xl p-4 text-xs overflow-x-auto"
         style={style}
       >
