+++ multimodal/tarko/ui/src/components/markdown-renderer/components/Image.tsx
