+++ multimodal/tarko/ui/src/components/markdown-renderer/styles/markdown.css
