+++ multimodal/tarko/ui/src/components/markdown-renderer/styles/syntax-highlight.css
