+++ multimodal/tarko/ui/src/components/markdown-renderer/utils/index.ts
