+++ multimodal/tarko/ui/src/components/markdown-renderer/utils/linkPreprocessor.ts
@@ -16,7 +16,7 @@
  * 4. Not already wrapped in markdown link syntax
  * 5. Followed by Chinese characters or other non-URL characters
  * 6. Excludes trailing punctuation that shouldn't be part of URLs
- * 
+ *
  * Updated to handle more edge cases and avoid false positives
  */
 const URL_REGEX =
@@ -32,13 +32,13 @@ export function preprocessMarkdownLinks(content: string): string {
   if (!hasBareUrlsWithChinese(content)) {
     return content;
   }
-  
+
   // Split content by code blocks to avoid processing URLs inside them
   const codeBlockRegex = /```[\s\S]*?```|`[^`]*`/g;
   const parts: Array<{ text: string; isCodeBlock: boolean }> = [];
   let lastIndex = 0;
   let match;
-  
+
   while ((match = codeBlockRegex.exec(content)) !== null) {
     // Add text before code block
     if (match.index > lastIndex) {
@@ -48,17 +48,15 @@ export function preprocessMarkdownLinks(content: string): string {
     parts.push({ text: match[0], isCodeBlock: true });
     lastIndex = match.index + match[0].length;
   }
-  
+
   // Add remaining text
   if (lastIndex < content.length) {
     parts.push({ text: content.slice(lastIndex), isCodeBlock: false });
   }
-  
+
   // Process only non-code-block parts
   return parts
-    .map(part => 
-      part.isCodeBlock ? part.text : part.text.replace(URL_REGEX, '[$1]($1)')
-    )
+    .map((part) => (part.isCodeBlock ? part.text : part.text.replace(URL_REGEX, '[$1]($1)')))
     .join('');
 }
