+++ multimodal/tarko/ui/src/index.ts
@@ -1,2 +1,3 @@
 export * from './components';
 export * from './utils';
+export * from './hooks';
